exceptions added

+++ Backend/symptom/src/main/java/com/sept_group6/symptom/controllers/SymptomController.java
@@ -2,6 +2,9 @@ package com.sept_group6.symptom.controllers;
 
 import com.sept_group6.symptom.dao.SymptomRepository;
 import com.sept_group6.symptom.model.Symptom;
+import com.sept_group6.symptom.exception.*;
+import org.springframework.transaction.TransactionSystemException;
+import javax.validation.ConstraintViolationException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,32 +15,82 @@ import java.util.*;
 @RestController
 @RequestMapping()
 public class SymptomController {
-    private static final Logger logger = LogManager.getLogger("Backend");
     @Autowired
     private SymptomRepository symptomRepository;
 
     @GetMapping("/getsymptom")
     public ResponseEntity<?> getSymptom(@RequestParam("email") String email) {
-        List<Symptom> symptom =
-                symptomRepository.findAllByEmail(email);
-        return ResponseEntity.accepted().body(symptom);
+        try {
+            if (!symptomRepository.existsByEmail(email)) {
+                throw new UserNotFoundException("There is no user with this email.");
+            }
+            List<Symptom> symptom = symptomRepository.findAllByEmail(email);
+            return ResponseEntity.accepted().body(symptom);
+        } catch (UserNotFoundException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        } catch (ConstraintViolationException e) {
+            String errorMessage = "";
+            for (var error : e.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            return ResponseEntity.badRequest().body(errorMessage);
+        } catch (TransactionSystemException e) {
+            ConstraintViolationException causeException = (ConstraintViolationException) GetRootException
+                    .getRootException(e);
+            String errorMessage = "";
+            for (var error : causeException.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            return ResponseEntity.badRequest().body(errorMessage);
+        }
     }
 
     @DeleteMapping("/deletesymptom")
     public void deleteSymptom(@RequestParam("id") int id) {
-        symptomRepository.deleteById(id);
+        try {
+            if (!symptomRepository.existsById(id)) {
+                throw new UserNotFoundException("There is no user with this id.");
+            }
+            symptomRepository.deleteById(id);
+        } catch (UserNotFoundException e) {
+            ResponseEntity.badRequest().body(e.getMessage());
+        } catch (ConstraintViolationException e) {
+            String errorMessage = "";
+            for (var error : e.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            ResponseEntity.badRequest().body(errorMessage);
+        } catch (TransactionSystemException e) {
+            ConstraintViolationException causeException = (ConstraintViolationException) GetRootException
+                    .getRootException(e);
+            String errorMessage = "";
+            for (var error : causeException.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            ResponseEntity.badRequest().body(errorMessage);
+        }
     }
 
-    @PutMapping(path = "/addsymptom", consumes="application/json", produces =
-            "application/json")
-    public ResponseEntity<?> addSymptom(@RequestBody Symptom newSymptom)
-            throws Exception {
-        logger.info(newSymptom);
-
-        // add resource
-        Symptom symptom = symptomRepository.save(newSymptom);
-
-        return ResponseEntity.accepted().body(symptom);
+    @PutMapping(path = "/addsymptom", consumes="application/json", produces = "application/json")
+    public ResponseEntity<?> addSymptom(@RequestBody Symptom newSymptom) {
+        try {
+            Symptom symptom = symptomRepository.save(newSymptom);
+            return ResponseEntity.accepted().body(symptom);
+        } catch (ConstraintViolationException e) {
+            String errorMessage = "";
+            for (var error : e.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            return ResponseEntity.badRequest().body(errorMessage);
+        } catch (TransactionSystemException e) {
+            ConstraintViolationException causeException = (ConstraintViolationException) GetRootException
+                    .getRootException(e);
+            String errorMessage = "";
+            for (var error : causeException.getConstraintViolations()) {
+                errorMessage += error.getMessage() + "\n";
+            }
+            return ResponseEntity.badRequest().body(errorMessage);
+        }
 
     }
 }
